pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.8",
    "jinja2>=3.1.5",
    "numpy==1.23.5",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
    "python-multipart>=0.0.20",
    "ruff>=0.9.5",
    "soundfile>=0.13.1",
    "uvicorn>=0.34.0",
]
